UHF-9214: Initial commit for refactoring linked events paragraph.
