+++ modules/helfi_react_search/helfi_react_search.services.yml
@@ -1,15 +1,19 @@
+parameters:
+  helfi_react_search.request_timeout: 60
 services:
+  _defaults:
+    autoconfigure: true
+    autowire: true
+
   logger.channel.helfi_react_search:
     parent: logger.channel_base
+    arguments: ['helfi_react_search']
+
+  helfi_react_search.api_client:
+    parent: helfi_api_base.api_client_base
     arguments:
-      - 'helfi_react_search'
-  helfi_react_search_linked_events:
-    class: Drupal\helfi_react_search\LinkedEvents
-    arguments:
-      - '@cache.default'
-      - '@http_client'
       - '@logger.channel.helfi_react_search'
-      - '@language_manager'
-  Drupal\helfi_react_search\EventSubscriber\ElasticIndexSubscriber:
-    tags:
-      - { name: 'event_subscriber' }
+      - { timeout: '%helfi_react_search.request_timeout%' }
+
+  Drupal\helfi_react_search\LinkedEvents: ~
+  Drupal\helfi_react_search\EventSubscriber\ElasticIndexSubscriber: ~
